Remove begin ... rescue block from notifications
the `Notifier#notify` call was wrapped in a `begin ... rescue`
block in order to fix tests.

In case you launch tests on a platform where `libnotify` is not
installed, you can apply the patch added with this commit.
Refer to the new `TROUBLESHOOTING.md` file to find fixes
in case you get errors.